+++ docs/README.md
@@ -2,7 +2,7 @@
 ## 게임 초기화
 - [ ] 각 Car에 이름 입력 받기
   - [x] `예외처리` : 빈 값이 입력된 경우
-  - [ ] `예외처리` : 이름이 5글자 초과한 경우
+  - [x] `예외처리` : 이름이 5글자 초과한 경우
   - [ ] `예외처리` : 쉼표로 구분 되지 않은 경우
 - [ ] 시도 횟수 입력 받기
   - [x] `예외처리` : 빈 값인 경우
