+++ src/main/java/racingcar/message/ViewMessage.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.message;
 
-public interface Message {
+public interface ViewMessage {
     String INPUT_CAR_NAMES = "경주할 자동차들의 이름을 입력하세요.(쉼표(,) 기준으로 이름구분)";
     String INPUT_TRIAL_COUNT = "시도할 회수입력하세요";
 }
