--- src/main/java/racingcar/view/OutputView.java
@@ -1,4 +1,0 @@
-package racingcar.view;
-
-public class OutputView {
-}
